rename themapaginatable to themapaginadataview
