+++ src/client/pages/Thema/Bodem/BodemThema.tsx
@@ -5,7 +5,7 @@ import { useBodemData } from './useBodemData.hook.tsx';
 import type { LoodMetingFrontend } from '../../../../server/services/bodem/types.ts';
 import { PageContentCell } from '../../../components/Page/Page.tsx';
 import { ThemaPagina } from '../../../components/Thema/ThemaPagina.tsx';
-import { ThemaPaginaTable } from '../../../components/Thema/ThemaPaginaTable.tsx';
+import { ThemaPaginaTable } from '../../../components/Thema/ThemaPaginaDataView.tsx';
 import { useHTMLDocumentTitle } from '../../../hooks/useHTMLDocumentTitle.ts';
 
 export function BodemThema() {
