+++ src/client/pages/Thema/Erfpacht/ErfpachtDossierDetail.tsx
@@ -8,7 +8,7 @@ import { useDossierDetailData as useDossierDetailData } from './DossierDetail/us
 import { CollapsiblePanel } from '../../../components/CollapsiblePanel/CollapsiblePanel.tsx';
 import { PageContentCell } from '../../../components/Page/Page.tsx';
 import { ThemaDetailPagina } from '../../../components/Thema/ThemaDetailPagina.tsx';
-import { ThemaPaginaTable } from '../../../components/Thema/ThemaPaginaTable.tsx';
+import { ThemaPaginaTable } from '../../../components/Thema/ThemaPaginaDataView.tsx';
 import { useHTMLDocumentTitle } from '../../../hooks/useHTMLDocumentTitle.ts';
 import { useAfisThemaData } from '../Afis/useAfisThemaData.hook.tsx';
