+++ src/client/pages/Thema/Horeca/HorecaThema.tsx
@@ -4,7 +4,7 @@ import { useHorecaThemaData } from './useHorecaThemaData.hook.ts';
 import type { HorecaVergunningFrontend } from '../../../../server/services/horeca/decos-zaken.ts';
 import { PageContentCell } from '../../../components/Page/Page.tsx';
 import { ThemaPagina } from '../../../components/Thema/ThemaPagina.tsx';
-import { ThemaPaginaTable } from '../../../components/Thema/ThemaPaginaTable.tsx';
+import { ThemaPaginaTable } from '../../../components/Thema/ThemaPaginaDataView.tsx';
 import { useHTMLDocumentTitle } from '../../../hooks/useHTMLDocumentTitle.ts';
 
 const pageContentTop = (
