+++ src/client/pages/Thema/Inkomen/InkomenThema.tsx
@@ -4,7 +4,7 @@ import { tableConfigSpecificaties } from './Inkomen-thema-config.ts';
 import { useInkomenThemaData } from './useInkomenThemaData.hook.ts';
 import { PageContentCell } from '../../../components/Page/Page.tsx';
 import { ThemaPagina } from '../../../components/Thema/ThemaPagina.tsx';
-import { ThemaPaginaTable } from '../../../components/Thema/ThemaPaginaTable.tsx';
+import { ThemaPaginaTable } from '../../../components/Thema/ThemaPaginaDataView.tsx';
 import { getFeedbackDetailsByTableConfig } from '../../../components/UserFeedback/UserFeedback.helpers.ts';
 import { useHTMLDocumentTitle } from '../../../hooks/useHTMLDocumentTitle.ts';
