+++ src/client/pages/Thema/Jeugd/JeugdThema.tsx
@@ -5,7 +5,7 @@ import type { LeerlingenvervoerVoorzieningFrontend } from '../../../../server/se
 import { PageContentCell } from '../../../components/Page/Page.tsx';
 import { ParagaphSuppressed } from '../../../components/ParagraphSuppressed/ParagraphSuppressed.tsx';
 import { ThemaPagina } from '../../../components/Thema/ThemaPagina.tsx';
-import { ThemaPaginaTable } from '../../../components/Thema/ThemaPaginaTable.tsx';
+import { ThemaPaginaTable } from '../../../components/Thema/ThemaPaginaDataView.tsx';
 import { useHTMLDocumentTitle } from '../../../hooks/useHTMLDocumentTitle.ts';
 import {
   WMO_HELPDESK_HREF_TEL_LINK,
